requirements.txt
fastapi
uvicorn
pyjwt
redis
requests
pydantic
python-dotenv
pydantic-settings